// webapp/src/utils/ct-types.d.ts
export type DateStringNullable = string | null;

/**
 * A simple timestamp in ISO format, e.g. '2022-10-19 12:00:00'
 */
export type DateTimeString = string;

/**
 * A simple timestamp in ISO format, e.g. '2022-10-19 12:00:00'
 */
export type DateTimeStringNullable = string | null;
// ...
export enum InvitationStatus {
    NOT_INVITED = 'not_invited',
    ACCEPTED = 'accepted',
    PENDING = 'pending',
}
// ...
export type MetaDataEntityId = {
    createdDate: ZuluDate;
    createdPerson: {
        id: number;
    };
    modifiedDate?: ZuluDate;
    modifiedPerson?: {
        id: number;
    };
};
// ...
export type Person = {
    acceptedsecurity?: DateStringNullable;
    addressAddition?: string;
    /**
     * This computed field contains the age of this person if the date of birth is visible. If this person already dead, the age is calculated until the date of death.
     */
    age?: number;
    baptisedBy?: string;
    birthName?: string;
    birthday?: DateStringNullable;
    birthplace?: string;
    campusId?: number;
    canChat?: boolean;
    chatActive?: boolean;
    city?: string;
    cmsUserId?: string;
    country?: string;
    dateOfBaptism?: DateStringNullable;
    dateOfBelonging?: DateStringNullable;
    dateOfDeath?: DateStringNullable;
    dateOfEntry?: DateTimeStringNullable;
    dateOfResign?: DateTimeStringNullable;
    /**
     * List of department IDs
     */
    departmentIds?: Array<number>;
    /**
     * Edit security level of the current person. The user can edit fields upto this level.
     */
    editSecurityLevelForPerson: number;
    /**
     * Primary email address of the person.
     */
    email?: string;
    emails?: Array<{
        contactLabelId: number;
        email: string;
        isDefault: boolean;
    }>;
    familyImageUrl?: string | null;
    familyStatusId?: number | null;
    fax?: string;
    firstContact?: DateTimeStringNullable;
    firstName?: string;
    growPathId?: number;
    guid: string;
    id: number;
    imageUrl?: string | null;
    invitationStatus?: InvitationStatus | null;
    isArchived?: boolean;
    /**
     * This computed field is true iff a date of death is set for this person.
     */
    isDead?: boolean;
    job?: string;
    lastEditedDate?: ZuluDateNullable;
    lastLogin?: ZuluDateNullable;
    lastName?: string;
    latitude?: number | null;
    latitudeLoose?: number | null;
    longitude?: number | null;
    longitudeLoose?: number | null;
    meta?: MetaDataEntityId;
    mobile?: string;
    nationalityId?: number | null;
    nickname?: string;
    /**
     * String with Optigem ID or empty string if no ID is set.
     */
    optigemId?: string;
    phonePrivate?: string;
    phoneWork?: string;
    placeOfBaptism?: string;
    /**
     * @deprecated
     */
    privacyPolicyAgreement?: {
        date: DateStringNullable;
        typeId: number | null;
        whoId: number | null;
    };
    privacyPolicyAgreementDate?: DateStringNullable;
    privacyPolicyAgreementTypeId?: number | null;
    privacyPolicyAgreementWhoId?: number | null;
    referredBy?: string;
    referredTo?: string;
    /**
     * Security level of the current person. The user sees fields upto this level.
     */
    securityLevelForPerson: number;
    sexId?: number | null;
    statusId?: number;
    street?: string;
    tags?: Array<Tag>;
    title?: string;
    weddingDate?: DateStringNullable;
    zip?: string;
};
// ...
export type Status = {
    /**
     * ID of status
     */
    id: number;
    /**
     * Flag if status is member of the church
     */
    isMember: boolean;
    /**
     * Flag if that status is searchable
     */
    isSearchable: boolean;
    /**
     * Status name
     */
    name: string;
    /**
     * Translated status name
     */
    nameTranslated: string;
    /**
     * Only persons with that securitylevel can edit/select/delete that status
     */
    securityLevelId: number;
    /**
     * Abbreviation of name.
     */
    shorty: string;
    /**
     * Used to sort all statuses
     */
    sortKey: number;
};
// ...
export type Tag = TagBase & {
    /**
     * @deprecated
     */
    modifiedAt: string;
    /**
     * @deprecated
     */
    modifiedBy: number;
    /**
     * @deprecated
     */
    modifiedDate: string;
    /**
     * @deprecated
     */
    modifiedPid: string;
};

export type TagBase = {
    color: CtColor;
    description: string;
    id: number;
    name: string;
};
// ...
export type ZuluDate = `${number}-${number}-${number}T${number}:${number}:${number}Z`;

/**
 * A timestamp in Zulu time format, e.g. '2022-10-19T12:00:00Z'
 *
 * @deprecated
 */
export type ZuluDateDeprecated = string;

/**
 * A timestamp in Zulu time format, e.g. '2022-10-19T12:00:00Z'
 */
export type ZuluDateNullable = `${number}-${number}-${number}T${number}:${number}:${number}Z` | null;